kernel: move thread stack macros to new header
kernel.h is enormous and this can be cleanly separated out.
